Add copyright notice and link to GitHub to ee.sh
When you install this with bpkg, the ee.sh script is copied somewhere on the target system and is isolated from the rest of the files. This gives the user a clue to figure out what this file is and does if they find it on their filesystem later on.